requirements.txt
Flask==2.3.3
google-generativeai==0.4.1
random